more warnings

+++ src/examples/time/time-warp-test.c
@@ -358,7 +358,7 @@ usage:
 	shared = setup_shared_var();
 
 	parent = getpid();
-	for(i = 1; i < tasks; i++) {
+	for(i = 1; i < (int)tasks; i++) {
 		if (!fork())
 			break;
 	}
